legacy/requirements.txt
accelerate==1.12.0
addict==2.4.0
agent-client-protocol==0.8.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
av==16.1.0
babel==2.18.0
backcall==0.2.0
beautifulsoup4==4.14.3
bleach==6.3.0
blis==1.3.3
cachetools==7.0.1
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
comm==0.2.3
confection==0.1.5
cryptography==46.0.3
csvw==3.7.0
ctranslate2==4.7.1
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.13
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
dlinfo==2.0.0
docopt==0.6.2
docutils==0.22.4
elevenlabs==2.37.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
espeakng-loader==0.2.4
eval_type_backport==0.3.1
executing==2.2.1
faster-whisper==1.2.1
fastjsonschema==2.21.2
filelock==3.24.3
flatbuffers==25.12.19
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
giturlparse==0.14.0
google-auth==2.48.0
googleapis-common-protos==1.72.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.5.0
idna==3.11
importlib_metadata==8.7.1
invoke==2.2.1
ipykernel==7.2.0
ipython==8.12.3
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
keyring==25.7.0
language-tags==1.2.0
linkify-it-py==2.0.3
loguru==0.7.3
Mako==1.3.10
markdown-it-py==4.0.0
markdownify==1.2.2
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mcp==1.26.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
meson==1.10.1
misaki==0.9.4
mistral-vibe==2.3.0
mistral_common==1.9.1
mistralai==1.12.4
mistune==3.2.0
more-itertools==10.8.0
mpmath==1.3.0
murmurhash==1.0.15
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
num2words==0.5.14
numpy==2.4.2
onnxruntime==1.24.2
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==25.0
pandocfilters==1.5.1
parso==0.8.6
pexpect==4.9.0
phonemizer-fork==3.3.2
pickleshare==0.7.5
pillow==12.1.1
platformdirs==4.9.2
preshed==3.0.12
prompt_toolkit==3.0.52
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycountry==26.2.16
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
pydantic_core==2.41.5
pygame==2.6.1
Pygments==2.19.2
PyJWT==2.11.0
pyparsing==3.3.2
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
pyzmq==27.1.0
rdflib==7.6.0
referencing==0.37.0
regex==2026.2.28
reportlab==4.4.10
requests==2.32.5
rfc3986==1.5.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
segments==2.3.0
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
smmap==5.0.2
sounddevice==0.5.5
soundfile==0.13.1
soupsieve==2.8.3
soxr==1.0.0
spacy==3.8.11
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
sse-starlette==3.2.0
stack-data==0.6.3
starlette==0.52.1
sympy==1.14.0
termcolor==3.3.0
textual==8.0.0
textual-speedups==0.2.1
thinc==8.3.10
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.2
tomli_w==1.2.0
torch==2.10.0+cu130
torchvision==0.25.0+cu130
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==5.2.0
tree-sitter==0.25.2
tree-sitter-bash==0.25.1
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uc-micro-py==1.0.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
wasabi==1.1.3
watchfiles==1.1.1
wcwidth==0.6.0
weasel==0.4.3
webencodings==0.5.1
websockets==16.0
win32_setctime==1.2.0
wrapt==2.1.1
yarg==0.1.9
zipp==3.23.0
zstandard==0.25.0
